fix:#101-신고하기 기능 유저 컨트롤러로 이동
