+++ src/main/java/com/damyo/alpha/api/user/controller/UserController.java
@@ -2,6 +2,7 @@ package com.damyo.alpha.api.user.controller;
 
 import com.damyo.alpha.api.auth.domain.UserDetailsImpl;
 import com.damyo.alpha.api.picture.service.S3ImageService;
+import com.damyo.alpha.api.smokingarea.service.SmokingAreaService;
 import com.damyo.alpha.api.user.controller.dto.UserResponse;
 import com.damyo.alpha.api.user.domain.User;
 import com.damyo.alpha.api.user.domain.UserRepository;
@@ -35,7 +36,7 @@ public class UserController {
 
     private final UserService userService;
     private final S3ImageService s3ImageService;
-    private final UserRepository userRepository;
+    private final SmokingAreaService smokingAreaService;
 
     @Operation(summary = "유저 정보 조회", description = "유저의 정보를 반환한다.")
     @ApiResponses(value = {
@@ -102,5 +103,16 @@ public class UserController {
     public ResponseEntity<?> deleteUser(@AuthenticationPrincipal UserDetailsImpl details) {
         userService.deleteUser(details.getId());
         return ResponseEntity.ok().body("회원 삭제 완료");
+    }
+
+    @GetMapping("/report/{smokingAreaId}")
+    @Operation(summary = "흡연구역 신고", description = "흡연구역이 존재하지 않을 때 유저가 보내는 요청")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "흡연구역 신고 성공"),
+            @ApiResponse(responseCode = "R101", description = "이미 신고한 흡연구역을 다시 신고한 경우")
+    })
+    public ResponseEntity<?> reportSmokingArea(@PathVariable String smokingAreaId, @AuthenticationPrincipal UserDetailsImpl details) {
+        smokingAreaService.reportSmokingArea(smokingAreaId, details.getId());
+        return ResponseEntity.ok().body("신고 완료");
     }
 }
